moved cuts_manager
