+++ pyodsp/alg/bm/cuts_manager.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 
 from pyomo.environ import ConcreteModel, Constraint
 
-from .cuts import Cut, FeasibilityCut, OptimalityCut
-from .params import BM_SLACK_TOLERANCE, BM_MAX_CUT_AGE, BM_CUT_SIM_TOLERANCE
+from ..cuts import Cut, FeasibilityCut, OptimalityCut
+from ..params import BM_SLACK_TOLERANCE, BM_MAX_CUT_AGE, BM_CUT_SIM_TOLERANCE
 
 
 @dataclass
