+++ pyodsp/dec/dd/alg_root_bm.py
@@ -13,7 +13,7 @@ from .mip_heuristic_root import IMipHeuristicRoot
 from pyodsp.alg.bm.bm import BundleMethod
 from pyodsp.alg.bm.pbm import ProximalBundleMethod
 from pyodsp.alg.cuts import CutList
-from pyodsp.alg.cuts_manager import CutInfo
+from pyodsp.alg.bm.cuts_manager import CutInfo
 from pyodsp.alg.params import BM_DUMMY_BOUND
 from pyodsp.solver.pyomo_solver import SolverConfig
